+++ functions/validators.php
@@ -1,6 +1,24 @@
 <?php
 
 /**
+ * Проверка, завершился ли аукцион, и если да, обновление победителя
+ */
+function handleEndedAuction(mysqli $dbConnection, int $lotId): void
+{
+    $lot = getLotById($dbConnection, $lotId);
+
+    // Проверяем, завершен ли аукцион
+    $isAuctionEnded = strtotime($lot['ended_at']) < time();
+    if ($isAuctionEnded) {
+        $winnerId = getWinnerIdFromRates($dbConnection, $lotId);
+
+        if ($winnerId) {
+            updateLotWinner($dbConnection, $lotId, $winnerId);
+        }
+    }
+}
+
+/**
  * Валидирует введенную ставку
  *
  * @param mixed $rateValue
